+++ test/raylib/test_runner.clj
@@ -59,6 +59,7 @@
                      raylib.scenes.bezier-test
                      raylib.scenes.fan-test
                      raylib.scenes.clipbox-test
+                     raylib.scroll-test
                      raylib.easings-test
                      raylib.scenes.clock-test
                      raylib.scenes.easings-test